Update run.py paths is scope of env. setup, use exec instead of import
Script setup_environment.py is run on every new platform or when the
directories change, so it is also just the right place to update path
in run.py, instead of running separate script.

Add run.py to gitignore, so the local paths will not get committed.

Run main script with exec() instead of using import - although less
secure, it is more appropriate in this case. The run.py script is just
a helper proxy in our case, and the actual main script shall be the
__main__ script.