robot: Adjust left/right camera positions (bounding boxes inside robot)
Move side cameras more to the center of the robot to prevent their
hit-boxes surpassing the main robot body hit-box. This consequently
requires to move them above the robot top PCB in order for them to
actually see anything and not be blocked by the robot housing.
Additionally, slightly tilt them downward, so that they can capture
whole wall even when near it.