robot: Move cameras lower to not obstruct the lidar view
If side cameras are above the robot, they always obstruct the lidar
view, mostly with their cable and connector.
Use the fact that sides of the robot model are "empty" and that cameras
use OpenGL to "see", and move them under the top housing of the robot,
at the top of the wheels - just low enough so that their FOV is not
obstructed by the top round robot housing.
This also removes the requirement for them to be tilted down, so make
them level.
Move them slightly inwards (decrease x pos) to increase their side FOV
as much as possible.