robot: Adjust colour_sensor position to track color in front
Previously, sensor was mounted at the bottom of the robot with the
desire of checking the color under the robot, as front as possible.
This caused sensor to be placed too forward and saw nothing, as it was
in the robot housing.
Now the sensor is positioned as it would be the front facing camera,
placed in the front robot housing opening and tilted just the right
amount downwards to check the color in the front of the robot. Actually,
this is the maximum possible tilt - because the sensor is still inside
the robot, tilting it down full 90 degree would again cause it to just
see the inside of the front plastic housing, not the actual floor in
front of the robot.